+++ src/Jute.jl
@@ -5,6 +5,7 @@ module Jute
 include("utils.jl")
 export pprint_time
 export rowmajor_product
+export with_output_capture
 
 include("options.jl")
 export build_run_options