+++ test/utils.test.jl
@@ -39,4 +39,49 @@ rowmajor_product_test = testcase() do
 end
 
 
+capture_all = testcase() do
+    ret, out = with_output_capture() do
+        println(STDOUT, "stdout 1")
+        println(STDERR, "stderr 1")
+        println(STDOUT, "stdout 2")
+        println(STDERR, "stderr 2")
+        1
+    end
+
+    @test ret == 1
+    @test out == "stdout 1\nstderr 1\nstdout 2\nstderr 2\n"
+end
+
+
+pass_through = testcase() do
+    with_output_capture() do
+        ret, out = with_output_capture(true) do
+            println(STDOUT, "stdout 1")
+            println(STDERR, "stderr 1")
+            1
+        end
+
+        @test ret == 1
+        @test out == ""
+    end
+end
+
+
+restore_on_error = testcase() do
+    try
+        with_output_capture() do
+            println(STDOUT, "stdout 1")
+            println(STDERR, "stderr 1")
+            error("error")
+            1
+        end
+    catch e
+        @test isa(e, ErrorException)
+    end
+
+    # TODO: how do we check that the output is back to normal?
+    # (besides eyeballing the test output, of course)
+end
+
+
 end